pyproject.toml
[project]
name = "ai-startup-analyst"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "easyocr>=1.7.2",
    "pillow>=11.3.0",
    "pymupdf>=1.26.4",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "requests>=2.32.5",
    "speechrecognition>=3.14.3",
]
